WIP: precommit workflow version swapped with old
