+++ server.py
@@ -64,8 +64,7 @@ from tools import (  # noqa: E402
     DebugIssueTool,
     ListModelsTool,
     PlannerTool,
-    Precommit,
-    PrecommitWorkflowTool,
+    PrecommitTool,
     RefactorTool,
     TestGenerationTool,
     ThinkDeepTool,
@@ -170,8 +169,7 @@ TOOLS = {
     "consensus": ConsensusTool(),  # Multi-model consensus for diverse perspectives on technical proposals
     "listmodels": ListModelsTool(),  # List all available AI models by provider
     "planner": PlannerTool(),  # Interactive sequential planner using workflow architecture
-    "precommit": Precommit(),  # Pre-commit validation of git changes
-    "precommitworkflow": PrecommitWorkflowTool(),  # Step-by-step pre-commit validation workflow
+    "precommit": PrecommitTool(),  # Step-by-step pre-commit validation workflow
     "testgen": TestGenerationTool(),  # Comprehensive test generation with edge case coverage
     "refactor": RefactorTool(),  # Intelligent code refactoring suggestions with precise line references
     "tracer": TracerTool(),  # Static call path prediction and control flow analysis
@@ -206,11 +204,6 @@ PROMPT_TEMPLATES = {
     },
     "precommit": {
         "name": "precommit",
-        "description": "Validate changes before committing",
-        "template": "Run precommit validation with {model}",
-    },
-    "precommitworkflow": {
-        "name": "precommitworkflow",
         "description": "Step-by-step pre-commit validation workflow",
         "template": "Start comprehensive pre-commit validation workflow with {model}",
     },
