+++ test_simulation_files/test_module.py
@@ -2,13 +2,11 @@
 Sample Python module for testing MCP conversation continuity
 """
 
-
 def fibonacci(n):
     """Calculate fibonacci number recursively"""
     if n <= 1:
         return n
-    return fibonacci(n - 1) + fibonacci(n - 2)
-
+    return fibonacci(n-1) + fibonacci(n-2)
 
 def factorial(n):
     """Calculate factorial iteratively"""
@@ -16,7 +14,6 @@ def factorial(n):
     for i in range(1, n + 1):
         result *= i
     return result
-
 
 class Calculator:
     """Simple calculator class"""
