+++ tests/test_line_numbers_integration.py
@@ -6,7 +6,7 @@ from tools.analyze import AnalyzeTool
 from tools.chat import ChatTool
 from tools.codereview import CodeReviewTool
 from tools.debug import DebugIssueTool
-from tools.precommit import Precommit
+from tools.precommit import PrecommitTool as Precommit
 from tools.refactor import RefactorTool
 from tools.testgen import TestGenerationTool
